Update async methods and refactor file handling
Refactored file handling to use asynchronous methods which improves performance. UpdateLinksForPackage method in SharedFolders class has been updated to return a Task. The code for moving all files and subdirectories within a source directory to destination directory has been added. In case destination contains a file with the same name, an exception is being thrown if the hashes of the files do not match.